Create Wrench-Mask-BLE_Unicorn-Hat-HD.ino
Modified the NRF52832 Version of the Mask to use the Pimoroni Unicorn Hat HD LED Matrices in place of the custom DotStar Matrices used previously. 

This is the initial version. The code appears to work more as expected but the patterns and brightness settings have not been corrected to match the new matrices.